chore: move Job.cpp and Job.hpp to managers package
