+++ src/managers/JobManager.hpp
@@ -4,15 +4,13 @@
 
 #pragma once
 
-#include "../Job.hpp"
+#include "Job.hpp"
 
 #include <unordered_map>
 #include <nlohmann/json.hpp>
 
 class JobManager {
 public:
-    ~JobManager();
-
     Job create_job(const int &pipeline_id, const nlohmann::json &job_body);
 
     void remove_job(const int &pipeline_id);
