+++ src/server/Server.cpp
@@ -1,5 +1,5 @@
 #include "Server.hpp"
-#include "../Job.hpp"
+#include "../managers/Job.hpp"
 
 #define HTTP_CREATED 201
